Finalize meals controller

+++ Controllers/MealsController.cs
@@ -12,7 +12,6 @@ namespace FoodDelivery.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
-    [Authorize(Roles = "CUSTOMER,RESTAURANT_OWNER")]
     public class MealsController : ControllerBase
     {
         private readonly FoodDeliveryContext _context;
@@ -24,38 +23,29 @@ namespace FoodDelivery.Controllers
 
         // GET: api/Meals/ForRestaurant/5
         [HttpGet("ForRestaurant/{id}")]
+        [Authorize(Roles = "CUSTOMER,RESTAURANT_OWNER")]
         public async Task<ActionResult<IEnumerable<Meal>>> GetMealsForRestaurant(long id)
         {
             var restaurant = await _context.Restaurants.FindAsync(id);
 
             if (restaurant == null)
-            {
                 return NotFound();
-            }
 
+            if (!User.IsInRole("CUSTOMER") && restaurant.OwnerUserId != Utilities.ExtractUserId(User))
+                return Unauthorized("Cannot access a restaurant you are not the owner of");
 
             return restaurant.Meals.ToList();
-        }
-
-        // GET: api/Meals/5
-        [HttpGet("{id}")]
-        public async Task<ActionResult<Meal>> GetMeal(long id)
-        {
-            var meal = await _context.Meals.FindAsync(id);
-
-            if (meal == null)
-            {
-                return NotFound();
-            }
-
-            return meal;
         }
 
         // PUT: api/Meals/5
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPut("{id}")]
+        [Authorize(Roles = "RESTAURANT_OWNER")]
         public async Task<IActionResult> PutMeal(long id, Meal meal)
         {
+            if (meal.Restaurant.OwnerUserId != Utilities.ExtractUserId(User))
+                return Unauthorized("Cannot access a restaurant you are not the owner of");
+
             if (id != meal.Id)
             {
                 return BadRequest();
@@ -85,8 +75,12 @@ namespace FoodDelivery.Controllers
         // POST: api/Meals
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPost]
+        [Authorize(Roles = "RESTAURANT_OWNER")]
         public async Task<ActionResult<Meal>> PostMeal(Meal meal)
         {
+            if (meal.Restaurant.OwnerUserId != Utilities.ExtractUserId(User))
+                return Unauthorized("Cannot access a restaurant you are not the owner of");
+
             _context.Meals.Add(meal);
             await _context.SaveChangesAsync();
 
@@ -95,13 +89,15 @@ namespace FoodDelivery.Controllers
 
         // DELETE: api/Meals/5
         [HttpDelete("{id}")]
+        [Authorize(Roles = "RESTAURANT_OWNER")]
         public async Task<IActionResult> DeleteMeal(long id)
         {
             var meal = await _context.Meals.FindAsync(id);
             if (meal == null)
-            {
                 return NotFound();
-            }
+
+            if (meal.Restaurant.OwnerUserId != Utilities.ExtractUserId(User))
+                return Unauthorized("Cannot access a restaurant you are not the owner of");
 
             _context.Meals.Remove(meal);
             await _context.SaveChangesAsync();
